Clean up dependencies
Dependencies were updated and lombok was removed because it was unnecessary,
it is an annotation processor that runs at compile-time by the java compiler,
it doesn't get used in the Kotlin compiler rendering it useless.
see #https://stackoverflow.com/questions/46027391/is-it-possible-to-use-lombok-with-kotlin

Cleaned up tests and added a few more.

In the future look for a better logging dependency or use one manually (currently not working as expected).