Merge of protocol-comments with proposal_comments
Entity protocol-comments has been deleted because it's now useless.
Organized the project package structure a litte better.
When commenting on a protocol, the protocol info and the comment are sent together
alongside the fields "validated" and "newValidation" which say wether to validate or override the previous validation.
This permits several more comments after validation and new validations to occur, while updating the necessary fields in the protocol entity.

Added triggers on update on the entities "proposal" and "proposal_comments"
 to maintain consistency and update "last_modified" fields to the time of the update.